Added analytics pages
