+++ components/analytics/stats/bar-chart.module.css
