+++ components/analytics/stats/bar-chart.tsx
