+++ components/analytics/stats/device-icon.tsx
