+++ components/analytics/stats/devices.tsx
