+++ components/analytics/stats/locations.tsx
