+++ pages/app/analytics.tsx
@@ -1,0 +1,10 @@
+import AppLayout from "@/components/layout/app";
+import Analytics from "@/components/analytics";
+
+export default function ProjectAnalytics() {
+  return (
+    <AppLayout>
+      <Analytics />
+    </AppLayout>
+  );
+}
